Rename Alert components
